Fixes specs and updates for 4.8
